Split Controllers & Add Static Folder
