+++ Dtos/AlbumWithMusicsArtistsDto.cs
@@ -3,6 +3,6 @@ namespace MusiCloud.Dtos;
 public class AlbumWithMusicsArtistsDto : AlbumDto
 {
     public List<ArtistDto> Artists { get; set; } = [];
-    public List<MusicDto> Musics { get; set; } = [];
+    public List<MusicWithArtistDto> Musics { get; set; } = [];
 
 }
