+++ Dtos/MusicDto.cs
@@ -5,4 +5,5 @@ public class MusicDto
     public Guid Id { get; set; }
     public string? Title { get; set; }
     public MetadataDto? Metadata { get; set; }
+    public uint Track { get; set; }
 }
