+++ Interface/IArtistService.cs
@@ -1,0 +1,8 @@
+using MusiCloud.Models;
+
+namespace MusiCloud.Interface;
+
+public interface IArtistService
+{
+    Task<Artist?> GetArtistAsync(Guid artistId);
+}
