+++ Interface/IFileProcessService.cs
@@ -1,11 +1,11 @@
-namespace MusiCloud.Services {
-    public interface IFileProcessService{
-        
-        // Task<bool> SaveAsync();
-        Task HandleFileDeletedAsync(string filePath);
-        Task HandleFileCreatedAsync(string filePath);
-        Task HandleFileRenamedAsync(string filePath, string oldPath);
-        Task InitializeAsync();
-        Task CleanupAsync();
-    }
+namespace MusiCloud.Interface;
+public interface IFileProcessService
+{
+
+    // Task<bool> SaveAsync();
+    Task HandleFileDeletedAsync(string filePath);
+    Task HandleFileCreatedAsync(string filePath);
+    Task HandleFileRenamedAsync(string filePath, string oldPath);
+    Task InitializeAsync();
+    Task CleanupAsync();
 }
