--- MusiCloud.slnx
@@ -1,3 +1,0 @@
-<Solution>
-  <Project Path="MusiCloud.csproj" />
-</Solution>
