+++ Program.cs
@@ -1,7 +1,9 @@
 using Microsoft.EntityFrameworkCore;
 using MusiCloud.Data;
 using MusiCloud.Services;
+using MusiCloud.Interface;
 using Scalar.AspNetCore;
+using Microsoft.Extensions.FileProviders;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -9,6 +11,7 @@ var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddAutoMapper(AppDomain.CurrentDomain.GetAssemblies());
+builder.Services.AddDirectoryBrowser();
 
 // Inject sqlite3 database service
 builder.Services.AddDbContext<MusiCloudDbContext>(options =>
@@ -16,8 +19,11 @@ builder.Services.AddDbContext<MusiCloudDbContext>(options =>
         ?? throw new NullReferenceException("Database ConnectionString Not Found!"))
 );
 
-// Inject MusiCloud service
+// Inject service
 builder.Services.AddScoped<IMusicService, MusicService>();
+builder.Services.AddScoped<IAlbumService, AlbumService>();
+builder.Services.AddScoped<IArtistService, ArtistService>();
+builder.Services.AddScoped<ISearchService, SearchService>();
 
 builder.Services.AddScoped<IFileProcessService, FileProcessService>();
 
@@ -40,7 +46,16 @@ if (app.Environment.IsDevelopment())
     });
 }
 
-app.UseHttpsRedirection();
+// app.UseHttpsRedirection();
+
+app.UseDefaultFiles();
+
+app.UseStaticFiles();
+
+app.AddCustomStaticFile(app.Configuration.GetValue<string>("CoverCache") ??
+    Path.Combine(Directory.GetCurrentDirectory(), "CoverCache"), "/api/cover");
+app.AddCustomStaticFile(app.Configuration.GetValue<string>("MusicFolder") ??
+        Path.Combine(Directory.GetCurrentDirectory(), "MusicFolder"), "/api/file");
 
 app.MapControllers();
