+++ appsettings.json
@@ -8,5 +8,7 @@
   "AllowedHosts": "*",
   "ConnectionStrings": {
     "MusiCloud": "Data Source=MusiCloud.db"
-  }
+  },
+  "MusicFolder":"D:\\Cache\\Music",
+  "CoverCache":"D:\\Cache\\Cover"
 }
